fix(variation): numpy.arange returns empty array
When there is no variation in the expected hours number, the stop will
have the same value as start, causing numpy to return an empty array and
to fail further generations.

On top of that, when the array only contains the same value, the
algorithm never matches the expected number of hours since it is always
going to cross the limit (the minimum generated total of hours is always
greather than the maximum expected).